Refactor client as a singleton
Fixes #7

This commit started its life as an attempt to change how the gem handles configuration, and then quickly pivoted to the author asking, "what is the expected usage of this gem?" as a means of interrogating how best to inject the needed configuration at the right time. The first step was analyzing what values need configuring and then exercising my imagination about how consumers might interact with the gem.

I determined that in a use case for the gem like we have in H2, we would most likely want to configure the client once up-front and then make API requests at will without needing to reconfigure or worry about whether there's a fresh connection or not. This is similar to how we designed dor-services-client, which is structured as a singleton class serving as a façade, an interface hiding the implementation details within the gem.

From there, I made a number of supporting changes:

* Add usage documentation to the README
* Add zeitwerk as a dependency for managing local dependencie (over explicit requires)
* Remove the `config` gem: configuration is injected via the `Globus::Client.configure` method and lives in memory as an OpenStruct instance
*

And I also included some extraneous changes that I thought were defensible to get in while the gem is still relatively young:

* Add Laura and Mike as gem authors
* Rename some implemention details within the `Endpoint` class that were very tightly coupled to H2 and Stanford
* Be more strict about what's in the public interface of the gem's classes