Add missing standardrb auto-fix update script
We are currently exploring using standardrb for our Ruby linting needs (instead of plain Rubocop), and part of using standardrb is using its excellent auto-fix functionality when the versions of rubocop and rubocop-performance change behind the scenes. This script was already present in Argo and should be present in every codebase using standardrb, such as globus_client.