Refactor endpoint to be less coupled to H2
Fixes #23

Globus transfer operations do not need intimate knowledge of H2 concepts, such as works and versions. Instead, we need a path to operate on. Leave it to consumers like H2 to construct their own relative paths.

Also, decouple the Identity class from needing to know about SUNet IDs and email addresses, and instead expect a working Globus user ID (which just so happens to be a Stanford email).