Reopen connection to avoid timeouts
We have seen Faraday::ConnectionFailed exceptions in long running
list_files() calls with thousdands of recursive API calls. These run
fine in development environments and on Stage, but not in Prod. While we
try to sort out if there is a firewall issue we can try reopening the
connection on each request.

For context see:

- https://github.com/sul-dlss/happy-heron/issues/3008
- https://github.com/sul-dlss/operations-tasks/issues/3306