Do not authenticate as part of client configuration
For the initial token, use a dummy value to avoid hitting any APIs during configuration, allowing the `TokenWrapper` to handle auto-magic token refreshing. Why not immediately get a valid token? Our apps commonly invoke client `.configure` methods in the initializer in all application environments, even those that are never expected to connect to production APIs, such as local development machines.

NOTE: `nil` and blank string cannot be used as dummy values here as they lead to a malformed request to be sent, which triggers an exception not rescued by `TokenWrapper`